+++ Services/ExcelService.cs
@@ -12,6 +12,7 @@ namespace MasteryTest3.Services
             List<OrderItem> items = new List<OrderItem>();
 
             using (var package = new ExcelPackage(file.OpenReadStream())) {
+
                 ExcelWorksheet sheet = package.Workbook.Worksheets[0];
 
                 int rowCount = sheet.Dimension.Rows;
@@ -21,7 +22,7 @@ namespace MasteryTest3.Services
                     {
                         product = sheet.Cells[row, 1].Value == null ? null : new (){ Id = Convert.ToInt32(NullSafeString(sheet.Cells[row, 1].Value)) },
                         name = NullSafeString(sheet.Cells[row, 2].Value),
-                        quantity = Convert.ToInt32(NullSafeString(sheet.Cells[row, 3].Value)),
+                        quantity = Convert.ToDouble(NullSafeString(sheet.Cells[row, 3].Value)),
                         unit = NullSafeString(sheet.Cells[row, 4].Value),
                         remark = NullSafeString(sheet.Cells[row, 5].Value),
                     });
